Add files via upload
Perhaps the busiest, of all the commits. This commit is for learning Inheritance. This folder covers a lot of different topics such as encapsulation and abstraction. Honestly, I think that is simply because as an OOP principle, it is paramount that such OOP tools be used for the creation of an inherited object, so it would make sense that abstraction or encapsulation is intended for the use of those objects.